Add files via upload
This is a magic eightball game containing the following randomized messages: "It is certain", "For sure!", "Yes definitely", "Most likely", "Probably", "Ask again later", "Better not tell you now", "Don't count on it", "My sources say no", "Hell nah", "Lame ahh question"
The question must contain a question mark at the end or else the user will need to repeat the question with a question mark at the end. 
The user is prompted to answer with a yes or y OR no or n if they want to shake the 8-ball, and yes or y will result in a randomized message and no or n will result in the program quitting.